Better support for setting header table size to 0
The workaround introduced by #17 makes a server use a default table size
when creating response headers, even if the client requested size 0.

This has the consequence that header entries are incorrectly added to
the table, and can therefore be referenced by successive responses,
causing the client to fail parsing the response.

This patch simply skips adding an entry to the table if it doesn't fit.