Generalize `Client`
The type of `run` is

```haskell
run :: ClientConfig -> Config -> Client a -> IO a
```

which looks reasonable enough. However, `Client` is a synonym for

```haskell
type Client a = (Request -> (Response -> IO a) -> IO a) -> IO a
```

The client is given a `sendRequest` function which it can use to send requests,
as long as it provides a way to process the response. That too is reasonable,
but the confusing part is the use of `a` in the type of `sendRequest`: why does
this have to be related to the final result of the `Client`? It seems to
suggest that perhaps all of these are collected in some way, but of course that
cannot be the case since (the result would have to be `IO [a]` or something).

Indeed, if we look at the code, there _is_ no relation; in this commit we
therefore generalize `Client` to

```haskell
type Client a = (forall b. Request -> (Response -> IO b) -> IO b) -> IO a
```

This makes more sense: we get a callback that we can call as often as we we
like; whatever type we get back when we process the response is the result of
that callback; and we then use this callback to produce the final result,
whatever it is.

This requires _no_ code changes in the library at all; just needs `RankNTypes`
enabled in a few modules.

**Backwards compatibility**: this is strictly a generalization of the existing
API, so existing code will continue to work; it just gets a more precise type,
and can do more. _However_, there is one limitation, which is why I had to
modify one of the tests in the test suite: where before we could have a
`[Client ()]`, that is no longer possible (as it would require
impredicativity). I suspect that this will not really be an issue, but I do not
know for sure.