Step 6/6: No need to wait to enqueue `Output`
Prior to this refactoring, it was necessary to delay enqueuing a new `Output`
until there was data available: without that delay, `outputOrEnqueueAgain`
would check the queue and the streaming window, realize that the queue was
empty, and then re-enqueue the `Output`, causing it to be handled out-of-order
(which is problematic, because streams must be opened in order). We can't
simply change `outputOrEnqueueAgain` _not_ to check the queue, because (_after_
we have sent the original headers), if there are no output available, we should
indeed output other streams first.

However, in the previous steps we changed `outputOrEnqueueAgain` to treat the
stream initialization (`OObj`) case special: don't check the queue nor the
window size, but send the headers immediately, and only _then_ check the rest.
This means that we no longer have to wait: when the stream is initialized, the
headers will be sent immediately, thereby establishing the stream, in the
appropriate order. The benefit to the _user_ is that the request is now
initialized before any data is sent, which may be important for some
applications.