Avoid race condition
This is a subtle one! There was a race between two thread managers: the
`System.TimeManager.Manager` and the `Network.HTTP2.H2.Manager` (which wraps
the former). The race condition could arise as follows:

1. Suppose an exception is raised in one of the background threads (e.g.,
   the frame sender).

2. Then `stopAfter` writes a `Stop` instruction the `H2.Manager` queue;
   the exception gets then rethrown.

3. If the server is run in the scope of `withManager` (as it is done in
   `Network.Run.TCP.Timeout.runTCPServerWithSocket` for example), the
   exception is caught again, causing `stopManager` to be called.

4. Now one of two things can happen /first/ (for each managed thread):

   a. `stopManager` might call the timeout action, which will throw
      `TimeoutThread` to the thread
      (because we registered the threads with `registerKillThread`)

   b. The `H2.Manager` will process the `Stop` instruction, and throw
      `KilledByHTTP2ThreadManager` to the thread.

   The choice is non-deterministic, depending on thread scheduling by the RTS.

5. After this, the _other_ exception is then thrown to _same_ thread
   (which might at this point be busy handling the _first_ exception).

So not only is the type of exception that is thrown non-deterministic, we also
have the possibility that the thread will be interrupted handling that first
exception by the second exception.

This commit fixes this by putting the `H2.Manager` in charge (it is aware of
both managers, after all). When `stopAfter` writes `Stop` to the `H2.Manager`
queue, it waits until that `Stop` instruction has been processed; this will
(after this commit) also ensure that the threads have been removed from the
timeout manager. This guarantees that the exception that is delivered (and the
_only_ exception that is delivered) is the `KilledByHTTP2ThreadManager` from
`http2`.