Fix treatment of async exceptions
In https://github.com/kazu-yamamoto/http2/pull/92 we added an exception handler
that was meant to catch _all_ exceptions (sync and async). This got changed in
https://github.com/kazu-yamamoto/http2/pull/114 (specifically,
https://github.com/kazu-yamamoto/http2/pull/114/commits/52a9619ba95b67d469205cb0dea546ada8489baa):
when we moved from `Control.Exception` to `UnliftIO.Exception`, we got a
different behaviour for `catch` and friends (see
https://github.com/well-typed/grapesy/issues/193#issuecomment-2238704595) for a
full list. This commit fixes some unintended consequences of this change.