Avoid unnecessary empty data frames at end of stream
Currently, if a stream terminates with an empty data frame, we will send that
empty data frame even if it is succeeded by trailers that are marked end of
stream. That empty data frame is unnecessary. We should only send it if the
payload length is non-zero and we are not sending trailers.