Document invariant on 'outputQ'
The 'outputQ' invariant is what guarantees we won't send @RST_STREAM@ frames
ahead of some still-enqueued data frames for a stream.

Additionally, the previous commit was unnecessarily tracking whether headers had
been sent on a stream to deal with a bug which we have now *actually* fixed, so
we no longer track whether headers are sent per-stream as we were.